+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Conform `@apollo/client` init to the v4 API to suppress warnings. Refs STCOR-992.
 * Provide `attach` to `<HotKeys>` to avoid its deprecated `findDOMNode` fallback. Refs STCOR-996.
 * Implement async `modules` loading infrastructure and update related components. Refs STCOR-976.
+* Listen for SessionReady events and update state to trigger re-render. Refs STCOR-977.
 
 ## [11.0.0](https://github.com/folio-org/stripes-core/tree/v11.0.0) (2025-02-24)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v10.2.0...v11.0.0)